### 2020-12-11 #2904 @Kevin-Mattheus-Moerman
@Isurwars thanks for this submission. Since the core functionality appears to be close to a 1000 lines of code I have flagged this submission for an additional scope check by the editorial board. In particular our board will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. This is standard for submissions of this size. We should get back to you within about a week.
