### 2021-01-18 Isurwars/Correlation#1 @awvwgk
Uncaught exception for invalid command line arguments
Running this project with invalid command line arguments like a string passed to the ``-a`` option results in an unhandled exception:

    Starting program: /home/awvwgk/projects/src/joss/Correlation/_build_debug/Correlation test/test_1/xSi.car -a str
    terminate called after throwing an instance of 'std::invalid_argument'
      what():  stof

    Program received signal SIGABRT, Aborted.
    0x00007ffff7ac5615 in raise () from /usr/lib/libc.so.6

Preferrably the exception is caught by the program and converted to a meaningful error message for the user to understand the problem with the input.
