### 2021-01-18 Isurwars/Correlation#2 @awvwgk
Too small atomic distances result in segmentation fault
Tested with a modified version of the test 1 input as shown here

    !BIOSYM archive 3
    PBC=ON
    Materials Studio Generated CAR File
    !DATE Thu Jul 30 00:37:03 2020
    PBC    5.4307    5.4307    5.4307   90.0000   90.0000   90.0000 (P1)
    Si1      0.000000000    0.000000000    0.000000000 XXXX 1      xx      Si  0.000
    Si2      0.000000000    2.715350000    2.715350000 XXXX 1      xx      Si  0.000
    Si3      2.715350000    0.000000000    2.715350000 XXXX 1      xx      Si  0.000
    Si4      2.715350000    2.715350000    0.000000000 XXXX 1      xx      Si  0.000
    Si5      4.073025000    1.357675000    4.073025000 XXXX 1      xx      Si  0.000
    Si6      1.357675000    1.357675000    1.357675000 XXXX 1      xx      Si  0.000
    Si7      1.357675000    4.073025000    4.073025000 XXXX 1      xx      Si  0.000
    Si8      1.357675000    4.073025000    4.073025000 XXXX 1      xx      Si  0.000
    end
    end

Note that atom 7 and 8 share the same cartesian coordinates here.
Running this results in a segmentation fault when creating the histogram

    Starting program: /home/awvwgk/projects/src/joss/Correlation/_build_debug/Correlation test/test_1/xSi.car
    File test/test_1/xSi.car opened successfully.
    [==================================================] 100 %
    Writing output files: test/test_1/xSi

    Program received signal SIGSEGV, Segmentation fault.
    Cell::PAD_Histogram (this=0x7fffffffcdb0, filename="test/test_1/xSi", theta_cut=180, bin_width=1) at ../src/Cell.cpp:640
    640	                        temp_hist[col][row]++;

I would expect the program to catch the too short distance before performing a calculation and report a meaningful error message regarding a potentially faulty input.
