### 2021-01-20 #2904 @awvwgk
I already briefly checked the project and reported issues while checking for the robustness of the user interface (see https://github.com/Isurwars/Correlation/issues/1 and https://github.com/Isurwars/Correlation/issues/2).
