### 2021-01-20 #2976 @whedon
[REVIEW]: Correlation: An Analyzing Tool for Liquids and for Amorphous Solids
**Submitting author:** @Isurwars (<a href="http://orcid.org/0000-0002-4359-2742">Isaías Rodríguez</a>)
**Repository:** <a href="https://github.com/Isurwars/Correlation/" target ="_blank">https://github.com/Isurwars/Correlation/</a>
**Version:** v1.0.2
