### 2021-01-20 c5a37aa
Initial argument and file validations as suggested by Sebastian Ehlert (@awvwgk)
