### 2021-01-20 Isurwars/Correlation#1 @Isurwars
Initial error catching in argument parsing has been added in the testing branch: https://github.com/Isurwars/Correlation/tree/Testing
