### 2021-01-20 Isurwars/Correlation#2 @Isurwars
Initial error handling in reading the file as well as distance has been incorporated.

Now when two atoms are closer than 0.5 Å (less than Bohr's radius) a warning is displayed but the process continious,
when two atoms are closer than 0.1 Å (10 pm) an error is raised and the code stops.

https://github.com/Isurwars/Correlation/tree/Testing
