### 2021-01-20 #2976 @Isurwars
An initial correction to invalid arguments as well as invalid position of atoms in the input file has been submited to the testing branch (https://github.com/Isurwars/Correlation/tree/Testing).

```
[=======================================>          ] 88 %
ERROR: The atoms:
Si_7 in position (4.07303, 4.07303, 1.35767),
Si_8 in position (4.11303, 4.07303, 1.35767),
Have a distance less than 10 pm.
```
```
[=======================================>          ] 88 %
WARNING: The atoms:
Si_7 in position (4.07303, 4.07303, 1.35767),
Si_8 in position (4.34302, 4.07303, 1.35767),
Have a distance less than the Bohr Radius.
[=======================================>          ] 98 %
Si_8 in position (4.34302, 4.07303, 1.35767),
Si_7 in position (4.07303, 4.07303, 1.35767),
Have a distance less than the Bohr Radius.
[==================================================] 100 %
Writing output files: test
```

How should i proceed?

Thanks for the feedback PhD. Sebastian Ehlert  (@awvwgk).
