### 2021-01-21 Isurwars/Correlation#2 @awvwgk
Thanks, this seems to resolve some issues, I found some strange behaviour with the `-w` option (see #3).
