### 2021-01-21 Isurwars/Correlation#1 @awvwgk
Thanks, this seems to resolve some of the issues with the command line interface. I encountered a strange issue with the `-w` option (see #3).
