### 2021-01-21 Isurwars/Correlation#4 @awvwgk
Joss Review
I'm looking into your Correlation project as part of your submission to JOSS (see https://github.com/openjournals/joss-reviews/issues/2976).

Some things I stumbled upon first looking at this project

- [x] ``clang++`` seems to be supported just fine, but is currently explicitly marked as not supported
- [x] the getting started section mentioned deployment, but there is no section on deployment
- [ ] the license file linked in the README is not present in the repository, the link should point to ``LICENSE`` rather than ``LICENSE.md``
- [x] binary files are checked into git (``bin/Correlation.exe``), I would recommend to remove them as they cannot be meaningfully tracked by a version control system
- [ ] a number of typos in the README makes it rather hard to read, especially if they present in program names (*e.g.* LAAMPS should be LAMMPS)
- [x] the test example in the README is not present in the repository at `test/test_1/aPd.cell` but `test/test_3/aPd.cell`

I will open separate issues as well for more concrete issues as well.
