### 2021-01-27 Isurwars/Correlation#2 @awvwgk
It would be more useful to catch those errors before starting the production run, checking for clashes should be doable without generating the periodic images. An easy solution would be to fold the cartesian coordinates to the central cell and compare the minimal image distance.

Also exiting at this point the application basically prohibits reusing the project, I would recommend to propagate the error back by throwing an exception.
