### 2021-01-27 Isurwars/Correlation#6 @awvwgk
Add contributing guidelines
Currently there is no clear guidance for contributors to learn about the projects. I usually consider it very helpful to have clear expectations spelled out in a dedicated contributing guideline. There are plenty of resources around for writing contributing guidelines, have a look at your favorite projects or maybe look at this collection here (https://github.com/nayafia/contributing-template) for inspiration.

If you don't want to add a full contributing guideline yet, consider at least to add a section to your README describing the recommended way to
1. submit patches
2. report bugs
3. ask for help

GitHub offers a community profile at every repository that can guide you through the process of refining your projects presentation to the open source community (see https://github.com/Isurwars/Correlation/community).
