### 2021-02-03 #2976 @awvwgk
I partly finished the review of the project in https://github.com/Isurwars/Correlation/issues/4 which lists inconsistencies in the documentation and basic blockers I found for this submission.

Also I reported a number of bugs in the command line interface (~API) of the project
- https://github.com/Isurwars/Correlation/issues/1
- https://github.com/Isurwars/Correlation/issues/2
- https://github.com/Isurwars/Correlation/issues/3
