### 2021-02-17 #2976 @CFGrote
> ## Review checklist for @CFGrote
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [ ]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/Isurwars/Correlation/)?
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [ ]  **Contribution and authorship:** Has the submitting author (@Isurwars) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> * [ ]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> ### Functionality
> * [ ]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
> * [ ]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [ ]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> * [ ]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> * [ ]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?
