### 2021-02-17 Isurwars/Correlation#7 @CFGrote
CFGrote's JOSS Review
Hi @Isurwars.
First off, great work and very useful. Please find my comments and notes while working on the JOSS Review checklist https://github.com/openjournals/joss-reviews/issues/2976.

#### Installation
* There is no `configure` step involved which means if I had non-standard settings on my computer, `make` might fail to locate the compiler and/or libraries. Also, installing into a target directory is not supported (no install target).
* You could explain why the user should run `make clean` **after** building.
* The binary produced by the `make all` step has the windows-typical '.exe' extension. This is a bit unusual for linux/mac.
#### Functionality
* `make tests` produces no errors, however it remains unclear if that just means the program ends or if it actually passed unittest and/or regression tests. Probably related to #5.
* Running the individual test with `./bin/correlation.exe test/test_1/aPd.cell`
  returns "File not found". aPd.cell is in test_3, test run is ok.
* Plotting the resulting PDF from test_3 gives the following figure:
![Figure_1](https://user-images.githubusercontent.com/15476807/108195232-ca784c00-7117-11eb-9684-64e887df2ac6.png).
 You might consider putting this or a similar graph into the README and instruct the user to try and reproduce this result as a confirmation the code works as it should.
* Performance: You claim that your code can handle up to 25000 atoms. I could not find a demonstration of such a system. Is it in test_6? If so, you should clearly mention it.

#### Documentation
* I find the output not properly documented (neither in the paper nor in the README): Describe the content of each file, describe what each column contains and specify the units. At a minimum, this information should go into the header of each .csv file. As a perspective, I recommend storing the results in a self-documenting format, e.g. hdf5.
* You should document the command line options in the README.
* No automated tests, see comment above. Adding tests that compare the code's output to reference data will make the validity of your code more accessible.
* No community guidelines (probably related to #6)

#### Paper
* Summary:
    * You state that the code is ready for mac while the README says mac not supported due to compiler issues.
    * A non-expert reader will be overwhelmed by the technical and field specific jargon. I advise to write something like: "Our code calculates various correlation functions from atomic positions".
    * Your claimed limit of 25000 atoms is not a lot (e.g. in biomolecules). What causes this limit? Does it depend on hardware/software on the user's system?
* Statement of need:
    * name and cite other packages (earlier, in statement of need). Please check the exhaustive list of software in the Computational Physics Communications archive http://cpc.cs.qub.ac.uk/, the IUCr's software list https://www.iucr.org/resources/other-directories/software for similar software. What is particular about your software that makes it stand out from other packages?
    * clearly name target audience
* Benchmarks, validation:
    * define test_x as being included in the software package. a table with
      references would be helpful.
    * Is test_6 really a challenge? it completes on my laptop in a few
      seconds. What are the limiting factors? memory/cpu? What exactly do you mean
      by "... benchmark  memory and CPU performance ...". The sentence on line 91-93 mentions test_6 and test_4, it is not clear which one you refer to here.
* Conclusion:
    * You mention HPC here but HTC (high throughput computing) earlier. What exactly do you mean here? Does your code support parallelization using e.g. MPI or similar frameworks? Suppose I want to run a large system that does not fit into memory, can your code deal with it? How do you address parallel IO?
    * How will it be extended, which features are missing?
