### 2021-02-17 #2976 @CFGrote
Review comments are in https://github.com/Isurwars/Correlation/issues/7.
