### 2021-02-23 Isurwars/Correlation#7 @jgostick
@CFGrote, this is superb review!  I think the authors of the package will be very appreciative of your efforts.
