### 2021-02-23 Isurwars/Correlation#8 @awvwgk
Software paper review
This is issue is part of the review in https://github.com/openjournals/joss-reviews/issues/2976, focusing on the software paper.

The paper itself is well-written and compares mainly against proprietary trajectory analyzers. The software is proposed as open source alternative to those proprietary programs. The paper does not compare to existing open source implementations for trajectory analysis. A non-comprehensive list of open source trajectory analyzers is given below:

- https://github.com/Amber-MD/cpptraj
- https://freud.readthedocs.io/en/stable/
- https://www.mdanalysis.org/
- https://mdtraj.org/1.9.4/index.html

Those projects should be mentioned and compared to this project.
