### 2021-02-23 #2976 @awvwgk
I finished my review of the software paper in https://github.com/Isurwars/Correlation/issues/8, the issues regarding the software / documentation / testing are mentioned here https://github.com/openjournals/joss-reviews/issues/2976#issuecomment-772875660.
