### 2021-02-24 Isurwars/Correlation#8 @Isurwars
I will incorporate the comparison within the next version of the paper/software.
