### 2021-03-04 afb675b
RC3:
Most of the requested modifications by the refeeres has been made to the 
code.
