+++ src/Makefile.in
@@ -86,7 +86,7 @@ POST_INSTALL = :
 NORMAL_UNINSTALL = :
 PRE_UNINSTALL = :
 POST_UNINSTALL = :
-bin_PROGRAMS = $(top_builddir)/bin/correlation$(EXEEXT)
+bin_PROGRAMS = $(BIN_DIR)/correlation$(EXEEXT)
 subdir = src
 ACLOCAL_M4 = $(top_srcdir)/aclocal.m4
 am__aclocal_m4_deps = $(top_srcdir)/configure.ac
@@ -98,18 +98,16 @@ CONFIG_CLEAN_FILES =
 CONFIG_CLEAN_VPATH_FILES =
 am__installdirs = "$(DESTDIR)$(bindir)"
 PROGRAMS = $(bin_PROGRAMS)
-am___top_builddir__bin_correlation_OBJECTS =  \
-	__top_builddir__bin_correlation-Atom.$(OBJEXT) \
-	__top_builddir__bin_correlation-Cell.$(OBJEXT) \
-	__top_builddir__bin_correlation-Constants.$(OBJEXT) \
-	__top_builddir__bin_correlation-ReadFiles.$(OBJEXT) \
-	__top_builddir__bin_correlation-main.$(OBJEXT)
-__top_builddir__bin_correlation_OBJECTS =  \
-	$(am___top_builddir__bin_correlation_OBJECTS)
-__top_builddir__bin_correlation_LDADD = $(LDADD)
-__top_builddir__bin_correlation_LINK = $(CXXLD) $(AM_CXXFLAGS) \
-	$(CXXFLAGS) $(__top_builddir__bin_correlation_LDFLAGS) \
-	$(LDFLAGS) -o $@
+am___BIN_DIR__correlation_OBJECTS =  \
+	__BIN_DIR__correlation-Atom.$(OBJEXT) \
+	__BIN_DIR__correlation-Cell.$(OBJEXT) \
+	__BIN_DIR__correlation-Constants.$(OBJEXT) \
+	__BIN_DIR__correlation-ReadFiles.$(OBJEXT) \
+	__BIN_DIR__correlation-main.$(OBJEXT)
+__BIN_DIR__correlation_OBJECTS = $(am___BIN_DIR__correlation_OBJECTS)
+__BIN_DIR__correlation_LDADD = $(LDADD)
+__BIN_DIR__correlation_LINK = $(CXXLD) $(AM_CXXFLAGS) $(CXXFLAGS) \
+	$(__BIN_DIR__correlation_LDFLAGS) $(LDFLAGS) -o $@
 am__dirstamp = $(am__leading_dot)dirstamp
 AM_V_P = $(am__v_P_@AM_V@)
 am__v_P_ = $(am__v_P_@AM_DEFAULT_V@)
@@ -126,12 +124,11 @@ am__v_at_1 =
 DEFAULT_INCLUDES = -I.@am__isrc@
 depcomp = $(SHELL) $(top_srcdir)/depcomp
 am__maybe_remake_depfiles = depfiles
-am__depfiles_remade =  \
-	./$(DEPDIR)/__top_builddir__bin_correlation-Atom.Po \
-	./$(DEPDIR)/__top_builddir__bin_correlation-Cell.Po \
-	./$(DEPDIR)/__top_builddir__bin_correlation-Constants.Po \
-	./$(DEPDIR)/__top_builddir__bin_correlation-ReadFiles.Po \
-	./$(DEPDIR)/__top_builddir__bin_correlation-main.Po
[…]

[…]
