### 2021-04-11 a4e5caa
Release Candidate 4:
Major revision to documentation required by JOSS editiorial.
