+++ src/main.cpp
@@ -66,7 +66,7 @@ void PrintHelp()
       "    HELP OPTIONS:"
       "      -h, --help\n"
       "        Display this help text.\n\n"
-      "    RADIAL OPTIONS:"
+      "    RADIAL OPTIONS:\n"
       "      -r, --r_cut\n"
       "        Cutoff radius in the calculation of g(r), G(r) and J(r). The default\n"
       "        radius it's set to 2 nm. The maximum recomended radius is the same as\n"
@@ -74,7 +74,7 @@ void PrintHelp()
       "        above this PBC value can be affected by periodic interactions.\n\n"
       "      -w, --bin_width\n"
       "        Width of the histograms for g(r) and J(r), the default is 0.05 nm.\n\n"
-      "    BOND-ANGLE OPTIONS:"
+      "    BOND-ANGLE OPTIONS:\n"
       "      -a, --angle_bin_width\n"
       "        Width of the histograms for the PAD, default set to 1.0°.\n\n"
       "      -b, --bond_parameter\n"
@@ -100,7 +100,7 @@ void PrintHelp()
       "             C  Mg 2.07\n"
       "        If any of the pairs is missing in the input file, the corresponding\n"
       "        bond distance will be set using the bond_parameter(1.30 by default).\n\n"
-      "    OUTPUT OPTIONS:"
+      "    OUTPUT OPTIONS:\n"
       "      -o, --out_file\n"
       "        The output file name, by default the input seed name will be used.\n\n"
       "CREATOR:\n"
