+++ src/Makefile.in
@@ -265,7 +265,7 @@ BIN_DIR = $(top_builddir)/bin
 __BIN_DIR__correlation_SOURCES = Atom.cpp Cell.cpp Constants.cpp ReadFiles.cpp main.cpp
 # correlation ld flags
 __BIN_DIR__correlation_CPPFLAGS = -Iinclude -std=c++17
-__BIN_DIR__correlation_LDFLAGS = -lm -static-libstdc++
+__BIN_DIR__correlation_LDFLAGS = -lm
 all: all-am
 
 .SUFFIXES:
