### 2021-07-05 #2976 @Isurwars
Sorry for the delay in the answer.

Installation
•	There is no configure step involved which means if I had non-standard settings on my computer, make might fail to locate the compiler and/or libraries. Also, installing into a target directory is not supported (no install target).

•	We make a configure script with the help of GNU autotools. The installation process s followed now by the standard ./configure, make, make install. The old Makefile is included in case of incompatibility with the new GNU autotools option.

•	You could explain why the user should run make clean after building.

•	The command make clean deletes object files created during the compilation process, and is not required to be run after compilation, it is used to keep a clean folder, and it is only recommended to clean the mess for GitHub upload. This info has been added to the README.md file.

•	The binary produced by the make all step has the windows-typical '.exe' extension. This is a bit unusual for linux/mac.

•	You are right. The ‘.app’ extension should be used in MacOS; as far as we know there is no standard extension for executables in linux, so no extension should be used. The new Makefile automatically adds the system extension to the executable.
Functionality
•	make tests produces no errors, however it remains unclear if that just means the program ends or if it actually passed unittest and/or regression tests. Probably related to #5.

•	This has been fixed in the testing branch. Currently the testing branch displays a “successfully calculated.” message at the end, as well as warnings and errors.

Running the individual test with ./bin/correlation.exe test/test_1/aPd.cell
returns "File not found". aPd.cell is in test_3, test run is ok.
•	Tests were rearranged to coincide with the paper numerals.

•	Plotting the resulting PDF from test_3 gives the following figure:
 .
You might consider putting this or a similar graph into the README and instruct the user to try and reproduce this result as a confirmation the code works as it should.
•	Thanks for the recommendation, we added images for the test and csv files.
•	Performance: You claim that your code can handle up to 25000 atoms. I could not find a demonstration of such a system. Is it in test_6? If so, you should clearly mention it.

•	This test was removed from the Release Candidate 1 (RC1) test samples because it required more than 15 minutes and more than 8 GB in RAM, after the optimization introduced in RC2, this test was reintroduced as test_7.
Documentation
•	I find the output not properly documented (neither in the paper nor in the README): Describe the content of each file, describe what each column contains and specify the units. At a minimum, this information should go into the header of each .csv file. As a perspective, I recommend storing the results in a self-documenting format, e.g. hdf5.

•	I agree. We have added documentation in the README.md file. The support for hdf5 file as well as for other formats should be added in a future release.

•	You should document the command line options in the README.

•	I agree. See above.

•	No automated tests, see comment above. Adding tests that compare the code's output to reference data will make the validity of your code more accessible.

•	A complete reworking of the Makefile and of the installation process was made in RC2 and RC3. The documentation of the new automated processes was included in RC4.

•	No community guidelines (probably related to #6)

•	Paper
•	Summary:
o	You state that the code is ready for mac while the README says mac not supported due to compiler issues.

o	Corrected in RC4.

o	A non-expert reader will be overwhelmed by the technical and field specific jargon. I advise to write something like: "Our code calculates various correlation functions from atomic positions".

o	We included your phrase in the abstract. I hope that is enough.

o	Your claimed limit of 25000 atoms is not a lot (e.g. in biomolecules). What causes this limit? Does it depend on hardware/software on the user's system?

o	This limit was caused by the algorithm used to calculate the correlations in the system. In RC1 we switched from an all vs all paradigm to a near-neighbors paradigm, effectively reducing the computing time by an order of magnitude and simultaneously the RAM requirements were reduced by two orders of magnitude.
This change has improved the code much, so now samples of 100,000+ atoms can be handled even in a laptop with 16 GB of RAM in less than 10 minutes instead of requiring 64+ GB of RAM and a Workstation.
•
•	Statement of need:
o	name and cite other packages (earlier, in statement of need). Please check the exhaustive list of software in the Computational Physics Communications archive http://cpc.cs.qub.ac.uk/, the IUCr's software list https://www.iucr.org/resources/other-directories/software for similar software. What is particular about your software that makes it stand out from other packages?clearly name target audience

•	Benchmarks, validation:
o	define test_x as being included in the software package. a table with
references would be helpful.

o	I agree; the table is included in the README.md file.

o	Is test_6 really a challenge? it completes on my laptop in a few
seconds. What are the limiting factors? memory/cpu? What exactly do you mean
by "... benchmark memory and CPU performance ...". The sentence on line 91-93 mentions test_6 and test_4, it is not clear which one you refer to here.

o	Test 7 is now included; it should require about 10 minutes and more than 4 GB of RAM to compute with the new algorithm in a regular laptop.

•	Conclusion:
o	You mention HPC here but HTC (high throughput computing) earlier. What exactly do you mean here? Does your code support parallelization using e.g. MPI or similar frameworks? Suppose I want to run a large system that does not fit into memory, can your code deal with it? How do you address parallel IO?

o	HPC vs HTC corrected. Currently the code does not support MPI parallelization; this will be a top priority in the future. My code cannot deal with a system that does not fit in to memory. My code is designed to work with several structures in parallel, as required by Machine Learning (ML) algorithms and because of the modular output the code can easily be integrated in an ML workflow.

o	How will it be extended, which features are missing?

o	The support for different kinds of output files like hdf5 standard is one of the top priorities in the near future.
o	Also, the inclusion of other correlation functions like Velocity Correlation Functions to further improve the analysis of liquids and phase transitions is the other top priority.
o	Finally, the parallelization of the main loop could further improve the code by switching to a “divide-and-conquer” paradigm.
