### 2021-07-08 #2976 @awvwgk
@Isurwars thanks for addressing our comments. Personally, I would prefer to discuss those points in the issue tracker of your projects as it allows to more clearly associate changes or patches with an already raised topic, not only for us, but also for your users, but we can also have this discussion here.
