### 2021-07-21 Isurwars/Correlation#13 @CFGrote
25000 atoms
In paper.md (branch Testing) you still mention 25000 atoms as the limit of what your code can handle but state in https://github.com/openjournals/joss-reviews/issues/2976#issuecomment-874199005 that it can now handle up to 100000 atoms.
