### 2021-07-21 Isurwars/Correlation#7 @CFGrote
I'll make new issues for each further comment, would appreciate if the discussion continues there.
- [ ] typo in test report (#10)
- [ ] typo in readme (#11)
- [ ] Output file documentation (units missing) (#12)
- [ ] 25000 atoms (#13)
- [ ] HPC vs. HTC (#14)
- [ ] Statement of need: Comment not addressed (#15)
- [ ] List of improvements into paper.md (#16)
