### 2021-07-21 #2976 @CFGrote
My 2nd round of review comments are here: https://github.com/Isurwars/Correlation/issues/7#issuecomment-884233704 with links to issues for each comment.
