### 2021-07-23 Isurwars/Correlation#13 @Isurwars
Corrected in RC5.

We include a 20,000+ atoms sample for testing, The ram limitation in my home (32GB) doesn't let me test with a larger sample.

Hope the next week we can finally return to the university to test the code with a larger sample in the university workstation with 128GB of RAM.
