### 2021-07-24 c32ec72
Release Candidate 5:
Addresing corrections for JOSS review process.
