+++ paper.bib
@@ -279,3 +279,31 @@
 	keywords = {Order-disorder models},
 	annote = {Includes index},
 }
+
+@article{roe_ptraj_2013,
+	title = {{PTRAJ} and {CPPTRAJ}: {Software} for {Processing} and {Analysis} of {Molecular} {Dynamics} {Trajectory} {Data}},
+	volume = {9},
+	issn = {1549-9618, 1549-9626},
+	shorttitle = {{PTRAJ} and {CPPTRAJ}},
+	url = {https://pubs.acs.org/doi/10.1021/ct400341p},
+	doi = {10.1021/ct400341p},
+	language = {en},
+	number = {7},
+	urldate = {2021-07-24},
+	journal = {J. Chem. Theory Comput.},
+	author = {Roe, Daniel R. and Cheatham, Thomas E.},
+	month = jul,
+	year = {2013},
+	pages = {3084--3095},
+	file = {Full Text:C\:\\Users\\Isurwars\\Zotero\\storage\\YVE4NQ4Y\\Roe and Cheatham - 2013 - PTRAJ and CPPTRAJ Software for Processing and Ana.pdf:application/pdf},
+}
+
+@article{VASPKIT-2021,
+title = {VASPKIT: A user-friendly interface facilitating high-throughput computing and analysis using VASP code},
+journal = {Computer Physics Communications},
+volume = {267},
+pages = {108033},
+year = {2021},
+doi = {https://doi.org/10.1016/j.cpc.2021.108033},
+author = {Vei Wang and Nan Xu and Jin-Cheng Liu and Gang Tang and Wen-Tong Geng},
+}
