### 2021-07-24 9de438c
Minor paper.md corrections
