### 2021-07-24 e298fbe
Fixing Figur 1 position.
