### 2021-07-24 cc09567
Figure 1 position.
