### 2021-07-24 582dfab
Perspective bullet points correction.
