### 2021-07-24 746effb
Minor typos in paper.md
