### 2021-07-25 9d19576
Structure Factor calculation initial release
