### 2021-07-25 #2976 @awvwgk
I updated the issues at the repo and closed everything that is already addressed. Some issues are still unaddressed or only partially resolved.
