### 2021-07-26 #2976 @jgostick
Hi @lsurwars, @CFGrote only has 2 'unchecked boxes' in his list above, and they both seem addressable.  Can you finish these off?
- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
