### 2021-07-26 #2976 @awvwgk
Looks good to me now. I have only minor thing left unaddressed, but those don't hinder moving the paper forward (still would be nice if they get addressed). The only outstanding issue in my opinion are the missing contributing guidelines for the project (see https://github.com/Isurwars/Correlation/issues/6).
