### 2021-07-28 Isurwars/Correlation#6 @Isurwars
Currently working on the contributing guidelines which will be published in RC6, I expected to publish RC6 tomorrow on July 28.
