+++ CONTRIBUTING.md
@@ -1,0 +1,50 @@
+[![Version]( https://img.shields.io/badge/version-1.0.0(RC6)-green)]( https://img.shields.io/badge/version-1.0.0(RC6)-green) [![License]( https://img.shields.io/badge/license-MIT-brightgreen)]( https://img.shields.io/badge/license-MIT-brightgreen)[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.0-4baaaa.svg)](code_of_conduct.md)
+
+#  Correlation: Contributing guidelines
+We want to make contributing to this project easy and transparent, whether it's:
+
+* Reporting a bug
+* Submitting a fix
+* Proposing new features
+* Discussing the current state of the code
+
+## Code of Conduct
+
+This project and everyone participating in it are governed by our [**code of conduct**](https://github.com/Isurwars/Correlation/blob/main/CODE_OF_CONDUCT.md) guidelines. By participating, you are expected to uphold this code. Please report unacceptable behavior.
+
+## How to contribute to **Correlation**
+
+### **Reporting a bug:**
+
+* **Please, make sure the bug has not been previously reported** by searching on GitHub under [Issues](https://github.com/Isurwars/Correlation/issues).
+
+* **If you're unable to find an open issue addressing the problem**, [open a new one](https://github.com/Isurwars/Correlation/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible: the **structure used**, as well as the **parameters**. In case the program does not match the expected results, please include a sample of the expected behavior that is not occurring.
+
+
+### **Submitting a fix:**
+
+In case you have created a patch for an existing bug in **Correlation**, follow these steps:
+
+- Open a new GitHub **pull request (PR)** with the patch.
+- Ensure the **PR** description includes the following :
+   * Clearly describes the **bug** and your solution.
+   * Include the relevant issue number if applicable.
+   * Include the changes to the [**Changelog**](https://github.com/Isurwars/Correlation/blob/main/ChangeLog) and your contribution to the [**AUTHORS**](https://github.com/Isurwars/Correlation/blob/main/AUTHORS) file.
+
+### **Proposing new features**
+
+* **Please, make sure the feature is not currently been developed** by searching on the GitHub [**branches**](https://github.com/Isurwars/Correlation/branches).
+* If the feature is currently under development, **you are heartily welcome** to contribute within the corresponding branch.
+
+* If the feature is not under development, but is included in the planned future features in the [**README**](https://github.com/Isurwars/Correlation/blob/main/README.md), you are encouraged to create a new branch.
+
+* If the intended feature is not included in the planned future features in the [**README**](https://github.com/Isurwars/Correlation/blob/main/README.md), please contact the developers listed in the [**AUTHORS**](https://github.com/Isurwars/Correlation/blob/main/AUTHORS) file.
+
+
+### **Discussing the current state of the code:**
+
+* For any questions about **Correlation** source code, please contact the developers listed in the [**AUTHORS**](https://github.com/Isurwars/Correlation/blob/main/AUTHORS) file.
+
+
+## License
+By contributing to **Correlation**, you agree that your contributions will be licensed under the MIT License.
