### 2021-08-02 Isurwars/Correlation#7 @Isurwars
Professor Valladares (Group leader) and I consider that editor and reviewer's suggestions improve the code and the paper enough to be considered as great contributors to the project.

How should I incorporate them in the paper.md and the AUTHORS files, as co-authors?
