### 2021-08-02 Isurwars/Correlation#7 @CFGrote
Thanks for your kind offer, @Isurwars. You are the author and should
take full credit for it. I only did my job as a reviewer. Feel free to
acknowledge this in the paper and/or repo, e.g. in a Acknowledgements
section, but I'd rather not become a co-author.
