+++ src/Cell.cpp
@@ -486,7 +486,7 @@ void Cell::RDF_Histogram(std::string filename, double r_cut, double bin_width)
      * w_ij is the weighting factor for the partial of G_ij
      * w_ij = (#atoms_i * #atoms_j)/total_number_atoms^2
      */
-    std::vector<double> w_ij(n_, 1.0);
+    std::vector<double> temp_w_ij(n_, 1.0);
 
     std::vector<std::vector<double> > temp_hist(n_, std::vector<double>(m_, 0));
     // Fill the r values of the histogram
@@ -528,18 +528,18 @@ void Cell::RDF_Histogram(std::string filename, double r_cut, double bin_width)
 
     std::ofstream out_file(filename + "_J.csv");
     std::setprecision(6);
-    out_file << std::setw(11) << "r (Å),";
+    out_file << std::setw(13) << "r (Å),";
     for (i = 0; i < n; i++) {
         for (j = i; j < n; j++) {
-            header = this->elements[i] + "-" + this->elements[j] + ",";
-            out_file << std::setw(11) << header;
+            header = this->elements[i] + "-" + this->elements[j] + " (1/Å),";
+            out_file << std::setw(13) << header;
         }
     }
     out_file << std::endl << std::fixed;
 
     for (i = 0; i < m_; i++) {
         for (j = 0; j < n_; j++) {
-            out_file << std::setw(10) << temp_hist[j][i] << ",";
+            out_file << std::setw(11) << temp_hist[j][i] << ",";
         }
         out_file << std::endl;
     }
@@ -561,20 +561,21 @@ void Cell::RDF_Histogram(std::string filename, double r_cut, double bin_width)
 
     std::ofstream out_file2(filename + "_g.csv");
     std::setprecision(6);
-    out_file2 << std::setw(11) << "r (Å),";
+    out_file2 << std::setw(13) << "r (Å),";
     for (i = 0; i < n; i++) {
         for (j = i; j < n; j++) {
-            header = this->elements[i] + "-" + this->elements[j] + ",";
-            out_file2 << std::setw(11) << header;
-            w_ij[i + j + 1] = 2.0 * this->element_numbers[i] * this->element_numbers[j] / ( num_atoms * num_atoms);
-            if (i == j) w_ij[i + j + 1] *= 0.5;
+            header = this->elements[i] + "-" + this->elements[j] + " (1/Å),";
+            out_file2 << std::setw(13) << header;
+            temp_w_ij[i + j + 1] = 2.0 * this->element_numbers[i] * this->element_numbers[j] / ( num_atoms * num_atoms);
+            if (i == j) temp_w_ij[i + j + 1] *= 0.5;
         }
     }
+    this->w_ij = temp_w_ij;
     out_file2 << std::endl << std::fixed;
 
     for (i = 0; i < m_; i++) {
         for (j = 0; j < n_; j++) {
-            out_file2 << std::setw(10) << temp_hist[j][i] << ",";
+            out_file2 << std::setw(11) << temp_hist[j][i] << ",";
         }
         out_file2 << std::endl;
     }
@@ -582,12 +583,16 @@ void Cell::RDF_Histogram(std::string filename, double r_cut, double bin_width)
 
     /*
      * We calclate G(r) with the definition:
-     * G(r) = 4 * pi * r * rho_0 * [g(r) - 1]
+     * G(r) = 4 * pi * r * rho_0 * [g(r) - 1];
+     *
+     * Weighted partials are calculated with:
+     * G_ij = 4 * pi * r * rho_0 * [g_ij(r) - w_ij]
      */
 
     for (col = 1; col < n_; col++) {
         for (row = 1; row < m_; row++) {
[…]
