### 2021-08-05 #2976 @jgostick
Hi @Isurwars, have you addressed the remaining 2 comments by @CFGrote?

- State of the field: Do the authors describe how this software compares to other commonly-used packages?
- References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
