### 2021-08-05 #2976 @awvwgk
> [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Has been addressed now with https://github.com/Isurwars/Correlation/commit/7da2042083119b88d5f2c6c7d6ec21b70fcca241
