### 2021-08-06 Isurwars/Correlation#7 @Isurwars
Understood, will reference the issues in my next comment, currently making the list.
