### 2021-08-09 Isurwars/Correlation#7 @Isurwars
Sorry for the delay in the answer, these has been a rough weeks in my personal life, and I have been in and out of the hospital.

I will answer the petitions and requests as soon as I can.
