### 2021-09-03 60a2a41
RC8 Added more options to structure factor and fixing missing request of JOSS reviewers.
