+++ src/main.cpp
@@ -43,7 +43,7 @@ std::string _bond_file_name_ = "";
 bool _bond_in_file_  = false;
 double _r_cut_       = 20.0;
 double _bin_w_       = 0.05;
-double _q_bin_w_     = 0.15707963;
+double _q_bin_w_     = 0.1570796326;
 double _bond_par_    = 1.3;
 double _angle_bin_w_ = 1.0;
 
@@ -75,6 +75,9 @@ void PrintHelp()
       "        above this PBC value can be affected by periodic interactions.\n\n"
       "      -w, --bin_width\n"
       "        Width of the histograms for g(r) and J(r), the default is 0.05 nm.\n\n"
+      "    STRUCTURE FACTOR OPTIONS:\n"
+      "      -q, --q_bin_width\n"
+      "        Width of the histograms for S(Q), the default is 0.157079 nm^()-1.)\n\n"
       "    BOND-ANGLE OPTIONS:\n"
       "      -a, --angle_bin_width\n"
       "        Width of the histograms for the PAD, default set to 1.0°.\n\n"
@@ -119,13 +122,14 @@ void PrintHelp()
 
 void ArgParser(int argc, char ** argv)
 {
-    const char * const short_opts = "a:b:hi:o:r:w:";
+    const char * const short_opts = "a:b:hi:o:q:r:w:";
     const option long_opts[]      = {
         { "angle_bin_width", required_argument, nullptr, 'a' },
         { "bond_parameter",  required_argument, nullptr, 'b' },
         { "help",            no_argument,       nullptr, 'h' },
         { "in_bond_file",    required_argument, nullptr, 'i' },
         { "out_file",        required_argument, nullptr, 'o' },
+        { "q_bin_width",     required_argument, nullptr, 'q' },
         { "r_cut",           required_argument, nullptr, 'r' },
         { "bin_width",       required_argument, nullptr, 'w' },
         { nullptr,           no_argument,       nullptr, 0   }
@@ -174,6 +178,19 @@ void ArgParser(int argc, char ** argv)
             case 'o': // -o or --out_file
                 _out_file_name_ = optarg;
                 break;
+            case 'q': // -q or --q_bin_width
+                try{
+                    _q_bin_w_ = std::stof(optarg);
+                }
+                catch (const std::exception& e) {
+                    std::cout << "Invalid input argument: '"
+                              << optarg
+                              << "' in q_bin_width parameter '-q' "
+                              << "(Real number expected)."
+                              << std::endl;
+                    exit(1);
+                }
+                break;
             case 'r': // -r ot --r_cut
                 try{
                     _r_cut_ = std::stof(optarg);
@@ -187,7 +204,7 @@ void ArgParser(int argc, char ** argv)
                     exit(1);
                 }
                 break;
-            case 'w': // -w or -- bin_width
+            case 'w': // -w or --bin_width
                 try{
                     _bin_w_ = std::stof(optarg);
                 }
@@ -295,7 +312,7 @@ int main(int argc, char ** argv)
      * The _r_cut_ parameter is the cutoff distance in r-space,
      * the _q_bin_w_ parameter is the bin width to be used in q-space.
      */
-    MyCell.SQ(_out_file_name_, _r_cut_, _q_bin_w_);
+    MyCell.SQ(_out_file_name_, _q_bin_w_, _bin_w_, _r_cut_);
 
     /*
      * This function uses the angles to calculate the PAD.
