### 2021-09-05 Isurwars/Correlation#13 @Isurwars
The paper.md has been corrected since RC5, will add a test file of 140,000+ atoms provided by Alejandro de Leon and Salvador Villarreal as private communication in RC9.

This test required at least an hour to complete, and more than 16GB of RAM were used in the process.
