+++ src/Cell.h
@@ -60,11 +60,11 @@ public:
     // Bond-Angle Calulation ()
     void PAD(bool = true);
     // RDF Histograms  (max distance between atoms, bin width)
-    void RDF_Histogram(std::string, double = 20.0, double = 0.05);
+    void RDF_Histogram(std::string, double = 20.0, double = 0.05, bool = false);
     // Nc Histograms ()
     void Nc_Histogram(std::string);
     // Structure Factor Calculation
-    void SQ(std::string, double = 0.1571, double = 0.05, double = 20.0);
+    void SQ(std::string, double = 0.1571, double = 0.05, double = 20.0, bool = false);
     // PAD Histograms ()
     void PAD_Histogram(std::string, double = 180.0, double = 1.0);
     // Read Only Lattive Vectors
