+++ Makefile.in
@@ -1163,6 +1163,15 @@ test7:
 	@echo " Stress Test successfully completed."
 	@echo "--------------------------------------------------"
 
+test8:
+	@echo "--------------------------------------------------"
+	@echo " Stress Test 2: (Amorphous Palladium Gold Hydrogen)"
+	@echo "--------------------------------------------------"
+	$(EXE) -r 40.0 $(TST_DIR)/test_8/aPdAuH-stress.car
+	@echo "--------------------------------------------------"
+	@echo " Stress Test 2 successfully completed."
+	@echo "--------------------------------------------------"
+
 ###################################################################
 ###################################################################
 clean-test:
