+++ src/main.cpp
@@ -73,10 +73,10 @@ void PrintHelp()
       "    RADIAL OPTIONS:\n"
       "      -n, --normalize"
       "        Used to switch between weighted partials (default), or normalize"
-      "        all the partials to 1 when r tends to inf."
+      "        all the partials to 1 when r tends to infinity"
       "      -r, --r_cut\n"
       "        Cutoff radius in the calculation of g(r), G(r) and J(r). The default\n"
-      "        radius it's set to 2 nm. The maximum recomended radius is the same as\n"
+      "        radius it's set to 2 nm. The maximum recommended radius is the same as\n"
       "        shortest length of the periodic boundary conditions (PBC), anything\n"
       "        above this PBC value can be affected by periodic interactions.\n\n"
       "      -w, --bin_width\n"
