### 2021-09-09 Isurwars/Correlation#7 @Isurwars
The last request has been adresss in RC8, and RC9. Hope we can move the paper foward.
