### 2021-09-09 #2976 @jgostick
wow, this one took a while, but it seems that it's ready to be accepted.  Both reviewers are fully satisfied.  (Note that @CFGrote and the author did most of their exchanges in the issues tracker of the official package).  I will begin moving this through the acceptance stages.  @Isurwars, you'll have a bit more to do, such as create a doi at zenodo.  I'll make a checklist below for you. Congratulations.
