+++ paper.md
@@ -11,23 +11,27 @@ authors:
     orcid: 0000-0002-4359-2742
     affiliation: 1
   - name: Renela M. Valladares
+    orcid: 0000-0002-3819-2226
     affiliation: 2
   - name: Alexander Valladares
     affiliation: 2
   - name: David Hinojosa-Romero
+    orcid: 0000-0002-8686-965X
     affiliation: 1
   - name: Ulises Santiago
+    orcid: 0000-0003-2259-5984
     affiliation: 3
   - name: Ariel A. Valladares
+    orcid: 0000-0003-1505-2244
     affiliation: 1
 affiliations:
- - name: Instituto de Investigaciones en Materiales, Universidad Nacional Autónoma de México
+ - name:  Departamento de Materia Condensada, Instituto de Investigaciones en Materiales, Universidad Nacional Autónoma de México, Coyoacán, Ciudad de México, México.
    index: 1
- - name: Facultad de Ciencias, Universidad Nacional Autónoma de México
+ - name:  Departamento de Física, Facultad de Ciencias, Universidad Nacional Autónoma de México, Coyoacán, Ciudad de México, México.
    index: 2
- - name: Department of Computational and Systems Biology, University of Pittsburgh
+ - name: Department of Computational and Systems Biology, University of Pittsburgh School of Medicine, Pittsburgh, Pennsylvania, USA.
    index: 3
-date: 30 November 2020
+date: 09 September 2021
 bibliography: paper.bib
 csl: aps.csl
 ---
@@ -66,9 +70,8 @@ The structure factor is one of the most useful tools in analyzing the scattering
 The Fourier transform of the scattering intensity given by the structure factor $S(Q)$, yields the pair distribution function (PDF) $g(r)$ defined by:
 
 \begin{equation}\label{eq:PDF}
-G(r) = 4\pi\rho_0 r(g(r)-1)=\frac{2}{\pi}\int_{0}^{\infty} Q[S(Q)-1]sin(Qr)dQ,
+g(r) - 1= \frac{1}{2 \pi^2 \rho_0 r}\int_{0}^{\infty} Q[S(Q)-1]sin(Qr)dQ,
 \end{equation}
-where $G(r)$ is the reduced pair distribution function (rPDF) which will be discussed next.
 
 The pair distribution function could also be seen like a distance map inside the material, the $g(r)$ function gives how feasible is finding two atoms separated by the distance ($r$) as can be seen in \autoref{fig:RDF} [@ziman_models_1979].
 
@@ -76,7 +79,7 @@ The PDF is normalized so that, as $r \to \infty$, $g(r) \to 1$, Also, as $r \to 
 
 ### Reduced pair distribution function $G(r)$
 
-One of the most widely used pair correlation function is the reduced pair distribution function. This is defined as $G(r) = 4\pi \rho_0 r (g(r)-1)$. From this definition, and the previously discussed tendency at large $r$ of the PDF, it's clear that the reduced pair distribution function (rPDF) oscillates around zero as $r \to \infty$. It also becomes evident that as $r \to 0$ (for $r$ smaller than the closest pair of atoms), the rPDF behaves like $-4\pi \rho_0$.
+One of the most widely used pair correlation function is the reduced pair distribution function (rPDF), also known in the literature as differential pair distribution function ($D(r)$). This function is defined as $G(r) = 4\pi \rho_0 r (g(r)-1)$. From this definition, and the previously discussed tendency at large $r$ of the PDF, it's clear that the reduced pair distribution function (rPDF) oscillates around zero as $r \to \infty$. It also becomes evident that as $r \to 0$ (for $r$ smaller than the closest pair of atoms), the rPDF behaves like $-4\pi \rho_0$.
 
[…]
