### 2021-09-10 0f8f422
Updated Zenodo DOI
Updated the Zenodo DOI to the final v1.0.0
