### 2021-09-10 #2976 @Isurwars
>     * [x]  Double check authors and affiliations (including ORCIDs)
The affiliations have been updated, and most of the ORCIDs have been added,  Professor Alexander Valladares does not have an ORCID.

>     * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

The version used for the paper is v1.0.0, https://github.com/Isurwars/Correlation/releases/tag/V1.0.0

>     * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.  Following [this guide](https://guides.github.com/activities/citable-code/) if needed.

The version has been archived in zenodo: https://zenodo.org/record/5499455
With DOI:  10.5281/zenodo.5499455

>     * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Author's affiliations and ORCIDs have been updated in the zenodo release.
