### 2021-09-10 a273b7d
Minor typos in the README file
