### 2021-09-11 2b3c24a
Update zenodo badge
Zenodo Badge updated to reflect the v1.0.1
