### 2021-09-11 #2976 @Isurwars
I have fixed the DOIs and created a new release v1.0.1:

https://github.com/Isurwars/Correlation/releases/tag/v1.0.1

I also create a corresponding zenodo release:

https://zenodo.org/record/5501541
