### 2021-09-14 0ba50dd
Adding JOSS badge to the README
