+++ README.md
@@ -1,21 +1,24 @@
 ![](Images/Banner.png)
+
 # **Correlation**: An Analysis Tool for Liquids and for Amorphous Solids
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5499455.svg)](https://doi.org/10.5281/zenodo.5499455)[![Version]( https://img.shields.io/badge/version-1.0.0-green)]( https://img.shields.io/badge/version-1.0.0-green) [![License]( https://img.shields.io/badge/license-MIT-brightgreen)]( https://img.shields.io/badge/license-MIT-brightgreen)[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.0-4baaaa.svg)](code_of_conduct.md)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5499455.svg)](https://doi.org/10.5281/zenodo.5499455)[![Version](https://img.shields.io/badge/version-1.0.1-green)](https://img.shields.io/badge/version-1.0.1-green) [![License](https://img.shields.io/badge/license-MIT-brightgreen)](https://img.shields.io/badge/license-MIT-brightgreen)[![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.0-4baaaa.svg)](code_of_conduct.md)[![status](https://joss.theoj.org/papers/7d5240007263a90d59d97ab464598991/status.svg)](https://joss.theoj.org/papers/7d5240007263a90d59d97ab464598991)
 
-**Correlation** is an analysis tool for correlation functions and correlation related properties of materials. In particular, for atomistic structure files of heavily used material simulation software like: DMoL3 (*.CAR), CASTEP(*.CELL), ONETEP(*.DAT), LAMMPS(*.XYZ),etc...
+**Correlation** is an analysis tool for correlation functions and correlation related properties of materials. In particular, for atomistic structure files of heavily used material simulation software like: DMoL3 (_.CAR), CASTEP(_.CELL), ONETEP(_.DAT), LAMMPS(_.XYZ),etc...
 
 This program calculates the main correlation functions of a material:
- - Radial Distribution Function (J(r)).
- - Pair Distribution Function (g(r)).
- - Reduced Pair Distribution Function (G(r)).
- - Coordination Number (CN).
- - Plane-Angle Distribution (PAD).
- - Structure Factor (S(Q)).
 
+- Radial Distribution Function (J(r)).
+- Pair Distribution Function (g(r)).
+- Reduced Pair Distribution Function (G(r)).
+- Coordination Number (CN).
+- Plane-Angle Distribution (PAD).
+- Structure Factor (S(Q)).
 
 ## Usage
+
 ---
+
 ```
     USAGE: correlation [OPTIONS] [input_file]
 
@@ -39,7 +42,7 @@ This program calculates the main correlation functions of a material:
             Cutoff radius in the calculation of g(r), G(r) and J(r). The default
             radius it's set to 2 nm. The maximum recommended radius is the same as
             the shortest length of the lattice parameters of the cell, anything
-            above this PBC value can be affected by periodic interactions.		
+            above this PBC value can be affected by periodic interactions.
           -w, --bin_width
             Width of the histograms for g(r) and J(r), the default is 0.05 nm.
 
@@ -80,6 +83,7 @@ This program calculates the main correlation functions of a material:
 ```
 
 ## Installation
+
 ---
 
 **Correlation** is written in C++17 and is designed to be compatible with GNU Compiler Collection (GCC) v7.0 or superior.
@@ -89,22 +93,29 @@ This program calculates the main correlation functions of a material:
[…]
