### 2021-09-16 02a706e
First attemp to correct small q.
for all q < 0.2 we aproximate Cosine by small angles, the aproximation 
was taken from INTRODUCTION TO
MODERN STATISTICAL
MECHANICS by David Chandler, section 7.5
