### 2021-09-16 #2976 @Kevin-Mattheus-Moerman
@Isurwars

To process this work for acceptance there are some minor last steps to work on:
- [ ] Please edit (you have to do this manually) the Zenodo archive DOI such that the title matches that of the paper: `Correlation: An Analysis Tool for Liquids and for Amorphous Solids`

I have read the paper and have some minor comments below:
- [ ] Check additional space in `(rdfpy, RadialDistributionFunction )` which should be `(rdfpy, RadialDistributionFunction)`
- [ ] Just a suggestion, not required: The mathematical notation for the sine function in equation 1 seems italic. Perhaps use `\sin` in Latex so it is rendered in non-italic like a function.  Again this is a rather minor suggestion.
- [ ] You may choose to use bold writing for software names e.g. **Correlation**, however it is more common in JOSS if they are written in code script e.g. `Correlation`. This is just a suggestion, not required.
