+++ src/Cell.h
@@ -36,6 +36,8 @@ public:
     std::vector<std::vector<double> > G;
     // Matrix of S(Q) Histograms
     std::vector<std::vector<double> > S;
+    // Matrix of XRD Histograms
+    std::vector<std::vector<double> > X;
     // Matrix of f(theta) Histograms
     std::vector<std::vector<double> > f_theta;
     // Volume of the cell
@@ -65,6 +67,8 @@ public:
     void Nc_Histogram(std::string);
     // Structure Factor Calculation
     void SQ(std::string, double = 0.1571, double = 0.05, double = 20.0, bool = false);
+    // XRD Calculation
+    void XRD(std::string, double = 1.5406, double = 5.0, double = 90.0, double = 1.0);
     // PAD Histograms ()
     void PAD_Histogram(std::string, double = 180.0, double = 1.0);
     // Read Only Lattive Vectors
