### 2021-09-17 5e36f17
v1.0.2
Minor fix to paper.md and README.md
