+++ README.md
@@ -1,10 +1,10 @@
 ![](Images/Banner.png)
 
-# **Correlation**: An Analysis Tool for Liquids and for Amorphous Solids
+# `Correlation`: An Analysis Tool for Liquids and for Amorphous Solids
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5499455.svg)](https://doi.org/10.5281/zenodo.5499455) [![Version](https://img.shields.io/badge/version-1.0.1-green)](https://img.shields.io/badge/version-1.0.1-green) [![License](https://img.shields.io/badge/license-MIT-brightgreen)](https://img.shields.io/badge/license-MIT-brightgreen) [![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.0-4baaaa.svg)](code_of_conduct.md) [![status](https://joss.theoj.org/papers/7d5240007263a90d59d97ab464598991/status.svg)](https://joss.theoj.org/papers/7d5240007263a90d59d97ab464598991)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5501541.svg)](https://doi.org/10.5281/zenodo.5501541) [![Version](https://img.shields.io/badge/version-1.0.1-green)](https://img.shields.io/badge/version-1.0.1-green) [![License](https://img.shields.io/badge/license-MIT-brightgreen)](https://img.shields.io/badge/license-MIT-brightgreen) [![Contributor Covenant](https://img.shields.io/badge/Contributor%20Covenant-2.0-4baaaa.svg)](code_of_conduct.md) [![status](https://joss.theoj.org/papers/7d5240007263a90d59d97ab464598991/status.svg)](https://joss.theoj.org/papers/7d5240007263a90d59d97ab464598991)
 
-**Correlation** is an analysis tool for correlation functions and correlation related properties of materials. In particular, for atomistic structure files of heavily used material simulation software like: DMoL3 (_.CAR), CASTEP(_.CELL), ONETEP(_.DAT), LAMMPS(_.XYZ),etc...
+`Correlation` is an analysis tool for correlation functions and correlation related properties of materials. In particular, for atomistic structure files of heavily used material simulation software like: DMoL3 (_.CAR), CASTEP(_.CELL), ONETEP(_.DAT), LAMMPS(_.XYZ),etc...
 
 This program calculates the main correlation functions of a material:
 
@@ -86,11 +86,11 @@ This program calculates the main correlation functions of a material:
 
 ---
 
-**Correlation** is written in C++17 and is designed to be compatible with GNU Compiler Collection (GCC) v7.0 or superior.
+`Correlation` is written in C++17 and is designed to be compatible with GNU Compiler Collection (GCC) v7.0 or superior.
 
 ### Prerequisites
 
-The following steps are required for the correct compilation of **Correlation** project.
+The following steps are required for the correct compilation of `Correlation` project.
 
 #### Windows
 
@@ -158,7 +158,7 @@ Installing GNU Autotools:
 brew install automake autoconf libtool
 ```
 
-### Installing **Correlation**
+### Installing `Correlation`
 
 ---
 
@@ -246,7 +246,7 @@ For example:
 
 ### Stress test
 
-A file is included as an stress test to **Correlation** in a particular machine, this test should take several minutes to execute and may require several GB of RAM to successfully compute the correlation functions.
+A file is included as an stress test to `Correlation` in a particular machine, this test should take several minutes to execute and may require several GB of RAM to successfully compute the correlation functions.
 
 These are the results for palladium deuteride reported by Rodríguez, _et al._ https://arxiv.org/abs/2012.02934
 
