### 2021-09-17 #2976 @Isurwars
* [x]  Please edit (you have to do this manually) the Zenodo archive DOI such that the title matches that of the paper: `Correlation: An Analysis Tool for Liquids and for Amorphous Solids`

Done.

* [x]  Check additional space in `(rdfpy, RadialDistributionFunction )` which should be `(rdfpy, RadialDistributionFunction)`

Removed space in the paper.md

* [x]  Just a suggestion, not required: The mathematical notation for the sine function in equation 1 seems italic. Perhaps use `\sin` in Latex so it is rendered in non-italic like a function.  Again this is a rather minor suggestion.

Changed from:
```
\begin{equation}\label{eq:PDF}
g(r) - 1= \frac{1}{2 \pi^2 \rho*0 r}\int*{0}^{\infty} Q[S(Q)-1]sin(Qr)dQ,
\end{equation}
```
To:

```
\begin{equation}\label{eq:PDF}
g(r) - 1= \frac{1}{2 \pi^2 \rho*0 r}\int*{0}^{\infty} Q[S(Q)-1]\sin(Qr)dQ,
\end{equation}
```

* [x]  You may choose to use bold writing for software names e.g. **Correlation**, however it is more common in JOSS if they are written in code script e.g. `Correlation`. This is just a suggestion, not required.

The corrections were added in v1.0.2: https://github.com/Isurwars/Correlation/releases/tag/v1.0.2

And the new zenodo archive: https://zenodo.org/record/5514113
